+++ src/xSTIR/cSTIR/cstir_p.cpp
@@ -923,6 +923,16 @@ sirf::cSTIR_ScatterEstimatorParameter(DataHandle* hp, const char* name)
 }
 
 void*
+sirf::cSTIR_setPoissonNoiseGeneratorParameter
+(const DataHandle *hp, const char* name, const DataHandle* hv)
+{
+    auto& obj = objectFromHandle<PoissonNoiseGenerator>(hp);
+    if (sirf::iequals(name, "seed"))
+        obj.seed(dataFromHandle<int>(hv));
+    return new DataHandle;
+}
+
+void*
 sirf::cSTIR_setGeneralisedObjectiveFunctionParameter
 (DataHandle* hp, const char* name, const DataHandle* hv)
 {