+++ CHANGES.md
@@ -2,10 +2,12 @@
 ## vx.x.x
 
 * SIRF/STIR
-  - `ScatterEstimation` has extra methods that allow setting masks for the tail-fitting
+  - `ScatterEstimation` has extra methods that allow setting masks for the tail-fitting.
   - `ImageData` has extra method to zoom image using information from a template image, `zoom_image_as_template`.
   - Error raised in `AcquisitionSensitivityModel.[un]normalise` methods applied to a read-only object.
   - `DataContainer.supports_array_view` to test for zero-copy compatibility.
+  - SIRF interfaces (C++ and Python) for STIR Poisson noise generation utilities provided.
+  - `ImageData` and `AcquisitionData` have `.asarray(copy=None)` (NumPy-like behaviour: default zero-copy if contiguous, fallback to deepcopy otherwise) via `__array_interface__`.
 
 ## v3.8.1